supp l'infirmier que lorsqu'on a le role d'admin

+++ AppHopital/src/Controller/DeleteInfirmierController.php
@@ -4,19 +4,29 @@ namespace App\Controller;
 
 use App\Service\TableauInfirmier;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 class DeleteInfirmierController extends AbstractController
 {
     #[Route('/delete/infirmier/{id}', name: 'delete_infirmier')]
-    public function index($id)
+    public function index(Request $request, $id)
     {
-        $manager = new TableauInfirmier;
-        $appelApi = $manager->DeleteInfirmier($id);
+        $session = $request->getSession();
+        $role = $session->get('role');
+        if ($role == 'ROLE_ADMIN'){
+            $manager = new TableauInfirmier;
+            $appelApi = $manager->DeleteInfirmier($id);
 
-        
-        return $this->redirectToRoute('login');
+
+            return $this->redirectToRoute('login');
+        }else{
+            return $this->render('access_denied/index.html.twig', [
+                'controller_name' => 'DeleteInfirmierController',
+                'error' => "Vous n'êtes pas autorisé à aller sur cette page"
+            ]);
+        }
         
 
     }
